attempt to further improve imager<->arm interaction
since image-import will decompress our remastered tar to build the actual
image from it we can pretty much do with compression what we want. so,
first we used minimal compression which is mostly fine for our purposes.
now we also make it rsyncable. according to the manpage this option should
be compatible with regular unzipping it does however add metadata to the
gzipped tar that rsync can use to only sync segments where the uncompressed
data actually changed. i.e. it allows efficient delta sync without having
to -z the tar via rsync (which would entail gzipping it on the image server
later) but using the actual .gz

additionally in the host remastering script use rsync with what appears
to be sane options to rsync the image into a tmpdir and then copy that
to the final destination to make it very explicit that the image is
actually synced