stop removing a whole pile of software in an attempt to make the image boot
with the present remastering the phone ends up at "this phone needs
restoring from a PC or service center." during the actual unpack/flash
step (i.e. after the phone reboots from the bootloader into the image for
the first time). a code search yielded nothing on this string so my
best guess right now is that it is coming from the recovery firmware that
is doing the unpack/flash. not knowing the code I have no clue what the
check is that makes it trip but assuming it is some general integrity check
of the image by shrinking the list of things we purge we should be able to
identify the relevant package(s) by process of elimination.

so. start this epic quest of debugging off by not explicitly removing
libs as they are the lowest in the stack and thus might be kept otherwise
and make things work.

also keep lightdm as it would be a corner stone of image integrity

also keep system-image-* as they sound pretty darn important and also would
pose a good thing to anchor integrity to

also do not install kf5 bits for the time being as the image doesn't boot
anyway it only needlessly increases the image size for now