do a minimal remaster to get to the bottom of the broken image
as mentioned int he previous commmit the image is broken for unknown
reasons. via adb shell and looking at logs I've identified that it fails to
set the phablet password which I have reproduced indeed fails. it is
however not clear to me why exactly it fails. the user itself is
weirdly weird though and only has a subuid from what I can tell but not
actually an entry in passwd. so going to continue shrinking the removal to
the point where the image is booting/nomore

disable everything except for language pack stripping. this should
still yield a working image, if so I am going to add more things to the
remove list until it stops booting again. such a waste of time -.-